refactor : FileView.jsx handler코드 fileHandlers폴더로 분리 (refs #28)
