+++ frontend/src/components/panels/fileHandlers/deleteHandlers.js
@@ -1,0 +1,9 @@
+import { deletePdf, deleteTextFile, setMemoAsNotSource } from '../../../../../backend/api/backend';
+
+const deleteHandlers = {
+  pdf: async (id) => await deletePdf(id),
+  txt: async (id) => await deleteTextFile(id),
+  memo: async (id) => { await setMemoAsNotSource(id); return true; },
+};
+
+export default deleteHandlers;
